% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as_numeric_factor}
\alias{as_numeric_factor}
\title{turn a factor with number levels into the numbers themselves}
\usage{
as_numeric_factor(x)
}
\arguments{
\item{x}{the factor to be converted}
}
\value{
numeric vector
}
\description{
Make a factor whose levels are numbers (e.g. SOA) into a vector
that contains these numbers not as factor levels, but as numerics.
}
\examples{
a <- factor( c( 1, 2, 1, 1, 1, 4, 7, 8 ) )
as.numeric( a )
as.numeric.factor( a )
a <- data.frame( col1 = c( 1, 2, 3 ), col2 = c( 3, 4, 5 ) )
}
\author{
http://stackoverflow.com/questions/3418128/how-to-convert-a-factor-to-an-integer-numeric-without-a-loss-of-information
}

