% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exclude}
\alias{exclude}
\title{Exclude extreme values from a vector}
\usage{
exclude(vector, upper.abs = NA, lower.abs = NA, upper.std = NA,
  lower.std = NA, explicit = FALSE)
}
\arguments{
\item{vector}{The vector containing the data that is to be pruned.}

\item{upper.abs}{An absolute upper cutoff value}

\item{lower.abs}{An absolute lower cutoff value}

\item{upper.std}{The factor by which the standard deviation of \emph{vector}
will be multiplied to obtain the upper cutoff value ( mean plus upper.std * std.dev )}

\item{lower.std}{Equivalent to upper.std for lower cutoff value}

\item{explicit}{Set to TRUE to get written info output from the funtion}
}
\value{
vector A vector that contains NA in all slots in which the upper or lower
bound were exceeded
}
\description{
Excludes high and low outliers from a vector.
Criterion for exclusion is either a fixed upper/lower value,
or a multiple of the standard deviation.
}
\examples{
v <- rnorm( 1000, 0, 1)
max( v )
v.clean <- exclude( v, upper.abs = 2 )
max( v.clean, na.rm = TRUE )

v.clean <- exclude( v, upper.abs = 2, lower.abs = -2)
max( v.clean, na.rm = TRUE )
min( v.clean, na.rm = TRUE )

v.clean <- exclude( v, upper.std = 2, lower.abs = -1, explicit = TRUE )
max( v.clean, na.rm = TRUE )
min( v.clean, na.rm = TRUE )
}

