% Generated by roxygen2 (4.0.2): do not edit by hand
\name{glmm_fitted}
\alias{glmm_fitted}
\title{glmm_fitted}
\usage{
glmm_fitted(model, diagnostics = FALSE)
}
\arguments{
\item{model}{An lmerMod or glmerMod object}

\item{diagnostics}{TRUE/FALSE for whether you want to see a plot for
RanFix and fitted values (these should be distributed around 0)}
}
\value{
Returns a data frame with columns fixedFit, fitted, RanFix
}
\description{
Extracts the fitted values from a lmerMod or glmerMod model.
In one data frame, it adds the fixed effects alone (as fixedFit,
equivalent to fixef() ),
fixed plus random (as fitted)
and their difference (as RanFix)
For a GLMM, it uses the inverse function to create values in the original
scale, rather than in the link scale. So for example, if you have used
binary data that you want to interpret as percentages with a log link,
then the inverse log is used here and you get "real" numbers back.
}
\author{
Tobias Heed, 2013-09-26
}

