% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inv_logit}
\alias{inv_logit}
\title{inv_logit}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{the logit that needs conversion back to original scale}
}
\value{
vector with inverse logit values
}
\description{
inverse logit
use e.g. to plot GLMM results in the original scale (e.g. p or %)
}
\examples{
a <- log( 0.2 / (1 - 0.2) )
inv_logit( a )
}
\author{
Tobias Heed
}

