% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mPrint}
\alias{mPrint}
\title{mPrint}
\usage{
mPrint(model)
}
\arguments{
\item{model}{An lmerMod or glmerMod object}
}
\value{
nothing
}
\description{
Prints the summary of an lmerMod or glmerMod plus:
which optimizer was used, how many iterations were allowed for
optimization, and any warnings that occurred.
}
\author{
Tobias Heed, 2013-09 or so
}

