% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select_col}
\alias{select_col}
\title{select_col}
\usage{
select_col(df, colName)
}
\arguments{
\item{df}{the data frame from which the column should be selected}

\item{colName}{the name (string) of the column}
}
\value{
vector with selected column
}
\description{
select a column of a data frame by giving a name as string
}
\examples{
a <- data.frame( col1 = c( 1, 2, 3 ), col2 = c( 3, 4, 5 ) )
select_col( a, "col1" )
s <- "col2"
select_col( a, s )
}
\author{
Tobias Heed
}

