% Generated by roxygen2 (4.0.2): do not edit by hand
\name{glmm_fitted}
\alias{glmm_fitted}
\title{extract fixed and random components for each trial of a GLMM data set}
\usage{
glmm_fitted(model, diagnostics = FALSE)
}
\arguments{
\item{model}{An lmerMod or glmerMod object}

\item{diagnostics}{TRUE/FALSE for whether you want to see a plot for
RanFix and fitted values (these should be distributed around 0)}
}
\value{
Returns a data frame with columns fixedFit, fitted, RanFix
}
\description{
Extracts the fitted values from a lmerMod or glmerMod model for
each line of the original data frame.
In one column, it adds the fixed effects alone (as fixedFit,
equivalent to fixef() ),
fixed plus random (as fitted)
and their difference (as RanFix)
For a GLMM, it uses the inverse function to create values in the original
scale, rather than in the link scale (e.g. inv logit for a logit link function).
So for example, if you have used
binary data that you want to interpret as percentages with a log link,
then the inverse log is used here and you get "real" numbers back.
The fix.inv.logit is equivalent to the marginal estimates of lsmeans.
lsmeans also gives an estimate for the confidence interval (could be implemented
here a well at some point...)
}
\author{
Tobias Heed, 2013-09-26
}

